Move debug-ui to platform folder
