+++ platform/debug-ui-api/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+    id("android-library-convention")
+}
+
+android {
+}
+
+dependencies {
+    implementation(platform(libs.kotlin.bom))
+}
