+++ platform/debug-ui-api/proguard-rules.pro
