+++ platform/debug-ui-api/src/main/AndroidManifest.xml
