+++ platform/debug-ui/src/debug/res/values/strings.xml
