+++ platform/debug-ui/src/release/kotlin/com/fuzzyfunctors/reductions/debug/ui/DebugDrawer.kt
@@ -1,12 +1,11 @@
 package com.fuzzyfunctors.reductions.debug.ui
 
 import android.app.Application
-
+import com.fuzzyfunctors.reductions.debug.ui.api.DebugUI
 object DebugDrawer : DebugUI {
     override fun initialize(application: Application) {
         // no op
     }
-
     override fun show() {
         // no op
     }
